Improve result type detection for rows containing NULL
When the row used to determine the types of the columns in a result
set contains a NULL value from SQLite, we try to determine the type
from the column referenced in the query.  If the result column does
not reference a column in a table, we give up and signal an error.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@246 dcde13d4-9b1b-0410-ac9e-ef07de68c835