Error out more gracefully when bad things happen in RS_sqlite_import
The transaction code was non-functional because the error routine
exited the function.  We now record the error message and then
decide whether to commit or not.

Note that we still are not properly handling SQLITE_SCHEMA errors,
but this patch at least allows the DB to be closed after errors
occur.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@260 dcde13d4-9b1b-0410-ac9e-ef07de68c835