When repreparing stmt, reassign res->drvResultSet in fetch
This is related to a segfault seen on lamb1.  When multiple
connections are accessing the same DB and the schema changes
in the middle of a query, we may need to reprepare the query.
I'm not sure if that should just invalidate the entire fetch
since you will start over?

This patch makes sure we keep track of the newly prepared
statement and avoids the segfault and error on lamb1.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@270 dcde13d4-9b1b-0410-ac9e-ef07de68c835