Improve eol handling in RS_sqlite_getline
Return NULL if EOF _or_ empty line (only eol).  This allows proper
behavior when input file is missing a trailing eol marker.

This patch also makes the eol handling more general so that
arbitrary strings should be able to be used as eol.  However,
the code that uses this calls readLines which restricts us to
'\n', '\r\n', or '\r'.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@273 dcde13d4-9b1b-0410-ac9e-ef07de68c835