Fix bug in NA handling for factors in bind.data
This patch resolves the issue by checking for NA_INTEGER when
converting factor to string.  But it suggests that we may want to
simplify matters by disallowing factors at this level of the code and
converting to character at a higher level.  We might also consider
issuing a warning since when the data is extracted from the db, the
vector will not be a factor.

Thanks to Thomas Lumley for the bug report.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@283 dcde13d4-9b1b-0410-ac9e-ef07de68c835