R-ification step 1
RSQLite no longer attempts to support S-Plus compatibility.  This patch
begins a cleanup series of removing "compatibility" code, which has
not been tested in a long time, in an effort to make the RSQLite code
base easier to maintain and more familiar to a developer experienced
with writing R package C code.

To begin, we use SEXP instead of s_object and this allows us not to
have to declare USE_RINTERNALS.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@317 dcde13d4-9b1b-0410-ac9e-ef07de68c835