Update to SQLite 3.4.0, use SQLite amalgamation
This patch updates the bundled SQLite version to 3.4.0 and introduces
a new mechanism of including the bundled SQLite code.  Instead of
building a separate SQLite lib and linking to it, we now use SQLite's
amalgamation file which puts all of SQLite into a single C file.  This
has an added benefit of simplifying the build process on Windows.



git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@330 dcde13d4-9b1b-0410-ac9e-ef07de68c835