Fix bug in createParameterBinding: convert logical to integer, not double
This patch resolves errors seen in dbWriteTable reported by Daniel
Wolff and Markus Neteler that occurs when a logical column is given as
the bind.data argument.  We were accidentally recognizing logical as
double by using R's isNumeric.  We now explicitly check for isLogical
and treat as integer.  And we use isReal instead of isNumeric.


git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@339 dcde13d4-9b1b-0410-ac9e-ef07de68c835