sqliteTransactionStatement (dbCommit) now uses try with silent = TRUE
Without this, it is impossible to suppress the error messages that an
application may otherwise be able to handle gracefully.  If you want
to see the error message after a failed commit, use dbGetException.

However, RSQLite code, such as dbWriteTable that calls dbCommit and
then dbRollback on failure will reset the exception stored for
dbGetException.  This change makes it more difficult to assess
failures in dbWriteTable.

git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@383 dcde13d4-9b1b-0410-ac9e-ef07de68c835