Fix handling of NULL values for BLOBs, modify sqliteDataType for BLOB
sqliteDataType which is called by dbDataType now maps R lists to
the SQLite BLOB type.  Previously, such columns would have mapped to
TEXT.  This allows dbWriteTable to be used to create a table when the
input data.frame contains a column that is a list of raw vectors
representing BLOBs.

git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@445 dcde13d4-9b1b-0410-ac9e-ef07de68c835