Fix bug when handling malformed bind.data in prepared queries
A missing parameter in bind.data was not being raised as an error and
the computation was continuing with invalid data.  This would get
reported as a STRING_ELT() can only be applied to a 'character
vector', not a 'NULL' error and was reported to lead to a segfault.

With this patch, an error is thrown if a named parameter is not found
in bind.data as originally intended.

Thanks to Andrew Piskorski for the bug report.

git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@459 dcde13d4-9b1b-0410-ac9e-ef07de68c835