Enhance sqliteDataType to use storage.mode
The storage mode of a data.frame column is now used as part of the
type detection. Prior to this patch, all vectors with class other than
numeric or logical were mapped to a TEXT column. This patch uses the
output of storage.mode to map to integer and double vectors to INTEGER
and REAL columns, respectively.  All other modes are mapped to a TEXT
column.

Detection of BLOBs was improved to treat objects as BLOBs if
data.class is "list" or "AsIs" and storage.mode is "list". The AsIs
heuristic is required since a common way to specify a list of raw
vectors in a data.frame is to use I().  Previously a less specific
test was used by checking only is.list() which would return TRUE for
lists regardless of assigned class. This patch requires the list to
have a data class of "list" or "AsIs".

git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@462 dcde13d4-9b1b-0410-ac9e-ef07de68c835