Increasee SQLITE_MAX_COLUMN to 30000, MAX_VARIABLE_NUMBER to 40000
The compile-time limits for maximum columns in a table has been
increased from the default value of 2000 to 30000. The maximum number
of parameters in a prepared query has been increased from 999 to
40000.

Use of wide tables is not encouraged. Databases with tables exceeding
2000 columns may not be usable without a custom-compiled SQLite
library.

See http://www.sqlite.org/limits.html for details on SQLite limits.

git-svn-id: https://hedgehog.fhcrc.org/compbio/r-dbi/trunk@465 dcde13d4-9b1b-0410-ac9e-ef07de68c835