Only export relevant symbols from shared library.
Currently the shared library are exporting all functions, some of
which collides with other DBI drivers, such as RPostgres.  Change
default symbol visibility to hidden and export only functions that
should be reachable from the R code, either via the R macro
attribute_visible (in C) or the Rcpp macro RcppExport (in C++).