Define `HAVE_USLEEP` compile time option
Without this, sqlite3 can only wait whole seconds when
waiting on a locked DB, which greatly increases the
chance that processes time out.

`usleep()` is not used on Windows, and it is present on
every Unix system. It is actually considered to be
obsolate, in favor of `nanosleep()`, but systems still
have it for compatibility.